rename: 최근 거래 관련 컴포넌트 네이밍 변경
