+++ src/components/screen/LedgerScreen/RecentTransactionOverviewBottomElement.tsx
@@ -1,7 +1,7 @@
 import {dummyTransactionDatas} from 'screens/Tab/TransactionScreen';
 import TransactionList from 'components/@common/TransactionList';
 
-export default function RecentTransactionBottomElement() {
+export default function RecentTransactionOverviewBottomElement() {
   return (
     <TransactionList
       variant={'recent'}
